requirements.txt

python-telegram-bot==21.11
apscheduler
sqlite3-binary
